Fix length parsing for WEBP chunk.
Currently, the length of the WEBP payload is assumed to be written
according to RIFF file format (i.e., the four bytes immediately
after the ID field).  That's not correct:

https://developers.google.com/speed/webp/docs/riff_container#webp_file_header

The length of the WEBP payload is the entire rest of the RIFF file.
This probably hasn't been breaking anything because the value which
is used (incorrectly) to determine the length is the fixed string
"VP8 ", which decodes to a very large number (1448097824, to be
exact).  So, this line:

var data = string.substr(offset + 4 + 4, len);

... will always get the entire rest of the file, unless the image is
REALLY REALLY big.